test: Distribute empty-state tests into per-construct files
The codebase organizes tests by SQL construct, so the #236 empty-state cases
belong as each construct's negative cases rather than in a policy-themed file.
Move the EmptyStatePolicyTests cases into ConditionIfTests, HavingTests,
AggregateFilterTests, UpdateTests, DeleteTests, Inner/LeftJoinTests, CaseTests,
MergeTests, and SelectTests, and delete EmptyStatePolicyTests. The empty-WHERE
case duplicated ConditionIfTests.ConditionIf_WhenConditionIsFalse and is dropped.
The policy itself stays documented in .claude/rules/guards-and-empty-states.md.

No behavior change; assertions are identical.
